Finalize the ReadMe
