+++ src/Avalonia.Samples/CustomControls/SnowflakesControlSample/Controls/SnowflakesControl.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
@@ -8,12 +7,8 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Data;
 using Avalonia.Media;
-using Avalonia.Platform;
 using Avalonia.Rendering;
-using Avalonia.Rendering.SceneGraph;
-using Avalonia.Skia;
 using Avalonia.Threading;
-using SkiaSharp;
 using SnowflakesControlSample.Models;
 
 namespace SnowflakesControlSample.Controls;
@@ -139,7 +134,10 @@ public class SnowflakesControl : Control, ICustomHitTest
                 snowFlake.Radius);
         }
 
-        // Bonus: Render the score hint if any available.
+        // Bonus 1: Use a custom renderer using Skia-API to display the total score
+        context.Custom(new ScoreRenderer(Bounds, $"Your score: {Score:N0}"));
+        
+        // Bonus 2: Render the score hint if any available.
         foreach (var scoreHint in _scoreHintsCollection.ToArray())
         {
             // If the game is running, move each flake to it's updated position
@@ -160,23 +158,6 @@ public class SnowflakesControl : Control, ICustomHitTest
             
             context.DrawText(formattedText, scoreHint.GetTopLeftForViewport(Bounds, new Size(formattedText.Width, formattedText.Height)));
         }
-
-        // Display user score
-        
-        // OPTION 1: Use formatted Text
-        //
-        // var formattedText =
-        //     new FormattedText(
-        //         "Your score: " + Score.ToString("N0"),
-        //         CultureInfo.InvariantCulture,
-        //         FlowDirection.LeftToRight,
-        //         Typeface.Default,
-        //         30,
-        //         Brushes.Goldenrod);
-        // formattedText.TextAlignment = TextAlignment.Right;
-        
-        // OPTION 2: Use a custom renderer using Skia-API
-        context.Custom(new ScoreRenderer(Bounds, $"Your score: {Score:N0}"));
 
         base.Render(context);
 
@@ -211,89 +192,5 @@ public class SnowflakesControl : Control, ICustomHitTest
         }
 
         return snowFlake != null;
-    }
-
-    /// <summary>
-    /// A helper-class to create a custom draw operation.
-    /// </summary>
-    private class ScoreRenderer : ICustomDrawOperation
-    {
-        public ScoreRenderer(Rect bounds, string text)
-        {
-            Bounds = bounds;
-            Text = text;
-        }
-
-        
-        /// <inheritdoc />
-        public bool Equals(ICustomDrawOperation? other)
-        {
-            // Equals is not used in this sample and thus will be always false. 
-            return false;
-        }
-        
-        /// <inheritdoc />
-        public void Dispose()
-        {
-            // Nothing to dispose
-        }
-
-        /// <inheritdoc />
-        public bool HitTest(Point p)
-        {
-            // The score shouldn't be hit-test visible
-            return false;
-        }
-
-        /// <inheritdoc />
-        public void Render(ImmediateDrawingContext context)
-        {
-            // Try to get the skia-feature.
-            var leaseFeature = context.TryGetFeature<ISkiaSharpApiLeaseFeature>();
-            
-            // In case we didn't find it, render the text with a fallback.
-            if (leaseFeature == null)
-            {
-                var glyphs = Text.Select(c => Typeface.Default.GlyphTypeface.GetGlyph(c)).ToArray();
-
-                var glyphRun = new GlyphRun(Typeface.Default.GlyphTypeface,
-                    20,
-                    Text.AsMemory(),
-                    glyphs,
-                    Bounds.TopRight - new Point(50, 50));
-                
-                context.DrawGlyphRun(Brushes.Goldenrod, glyphRun.TryCreateImmutableGlyphRunReference()!);
-            }
-            // Otherwise use SkiaSharp to render the text and apply some glow-effect.
-            // Find the SkiaSharp-API here: https://learn.microsoft.com/en-us/dotnet/api/skiasharp?view=skiasharp-2.88 
-            else
-            {
-                using var lease = leaseFeature.Lease();
-                var canvas = lease.SkCanvas;
-                canvas.Save();
-                
-                using (var paint = new SKPaint())
-                {
-                    paint.Shader = SKShader.CreateColor(SKColors.Goldenrod);
-                    paint.TextSize = 30;
-                    paint.TextAlign = SKTextAlign.Right;
-                    
-                    var origin = Bounds.TopRight.ToSKPoint();
-                    origin.Offset(-25, +50);
-
-                    paint.ImageFilter = SKImageFilter.CreateDropShadow(0, 0, 10, 10, SKColors.White);
-                    canvas.DrawText(Text, origin, paint);
-                }
-                canvas.Restore();
-            }
-        }
-
-        /// <inheritdoc />
-        public Rect Bounds { get; }
-        
-        /// <summary>
-        /// Gets the Text to display.
-        /// </summary>
-        public string Text { get; }
     }
 }
